feat: bridge data and actions lane
